+++ src/datatype.c
@@ -52,7 +52,6 @@ JL_DLLEXPORT jl_methtable_t *jl_new_method_table(jl_sym_t *name, jl_module_t *mo
     jl_atomic_store_relaxed(&mt->leafcache, (jl_array_t*)jl_an_empty_vec_any);
     jl_atomic_store_relaxed(&mt->cache, jl_nothing);
     mt->max_args = 0;
-    mt->kwsorter = NULL;
     mt->backedges = NULL;
     JL_MUTEX_INIT(&mt->writelock);
     mt->offs = 0;
@@ -101,6 +100,7 @@ jl_datatype_t *jl_new_uninitialized_datatype(void)
     t->hasfreetypevars = 0;
     t->isdispatchtuple = 0;
     t->isbitstype = 0;
+    t->isprimitivetype = 0;
     t->zeroinit = 0;
     t->has_concrete_subtype = 1;
     t->cached_by_hash = 0;
@@ -170,7 +170,8 @@ HTIMPL_R(layoutcache, _hash_layout_djb2, layout_eq)
 static htable_t layoutcache;
 static int layoutcache_initialized = 0;
 
-static jl_datatype_layout_t *jl_get_layout(uint32_t nfields,
+static jl_datatype_layout_t *jl_get_layout(uint32_t sz,
+                                           uint32_t nfields,
                                            uint32_t npointers,
                                            uint32_t alignment,
                                            int haspadding,
@@ -213,6 +214,7 @@ static jl_datatype_layout_t *jl_get_layout(uint32_t nfields,
     jl_datatype_layout_t *allocamem = (jl_datatype_layout_t *)(should_malloc ? NULL : alloca(flddesc_sz));
     jl_datatype_layout_t *flddesc = should_malloc ? mallocmem : allocamem;
     assert(flddesc);
+    flddesc->size = sz;
     flddesc->nfields = nfields;
     flddesc->alignment = alignment;
     flddesc->haspadding = haspadding;
@@ -454,7 +456,6 @@ void jl_compute_field_offsets(jl_datatype_t *st)
     if (st == w && st->layout) {
         // this check allows us to force re-computation of the layout for some types during init
         st->layout = NULL;
-        st->size = 0;
         st->zeroinit = 0;
         st->has_concrete_subtype = 1;
     }
@@ -463,7 +464,6 @@ void jl_compute_field_offsets(jl_datatype_t *st)
     // and reused by all subtypes.
     if (w->layout) {
         st->layout = w->layout;
-        st->size = w->size;
         st->zeroinit = w->zeroinit;
         st->has_concrete_subtype = w->has_concrete_subtype;
         if (!jl_is_layout_opaque(st->layout)) { // e.g. jl_array_typename
@@ -479,18 +479,18 @@ void jl_compute_field_offsets(jl_datatype_t *st)
         // if we have no fields, we can trivially skip the rest
         if (st == jl_symbol_type || st == jl_string_type) {
             // opaque layout - heap-allocated blob
-            static const jl_datatype_layout_t opaque_byte_layout = {0, 1, -1, 1, 0, 0};
+            static const jl_datatype_layout_t opaque_byte_layout = {0, 0, 1, -1, 1, 0, 0};
             st->layout = &opaque_byte_layout;
             return;
         }
         else if (st == jl_simplevector_type || st == jl_module_type || st->name == jl_array_typename) {
-            static const jl_datatype_layout_t opaque_ptr_layout = {0, 1, -1, sizeof(void*), 0, 0};
+            static const jl_datatype_layout_t opaque_ptr_layout = {0, 0, 1, -1, sizeof(void*), 0, 0};
             st->layout = &opaque_ptr_layout;
             return;
         }
         else {
             // reuse the same layout for all singletons
-            static const jl_datatype_layout_t singleton_layout = {0, 0, -1, 1, 0, 0};
+            static const jl_datatype_layout_t singleton_layout = {0, 0, 0, -1, 1, 0, 0};
             st->layout = &singleton_layout;
         }
     }
@@ -611,9 +611,10 @@ void jl_compute_field_offsets(jl_datatype_t *st)
             if (al > alignm)
                 alignm = al;
         }
-        st->size = LLT_ALIGN(sz, alignm);
-        if (st->size > sz)
+        if (LLT_ALIGN(sz, alignm) > sz) {
             haspadding = 1;
+            sz = LLT_ALIGN(sz, alignm);
+        }
         if (should_malloc && npointers)
             pointers = (uint32_t*)malloc_s(npointers * sizeof(uint32_t));
         else
@@ -632,7 +633,7 @@ void jl_compute_field_offsets(jl_datatype_t *st)
             }
         }
         assert(ptr_i == npointers);
-        st->layout = jl_get_layout(nfields, npointers, alignm, haspadding, desc, pointers);
+        st->layout = jl_get_layout(sz, nfields, npointers, alignm, haspadding, desc, pointers);
         if (should_malloc) {
             free(desc);
             if (npointers)
@@ -680,7 +681,6 @@ JL_DLLEXPORT jl_datatype_t *jl_new_datatype(
     jl_gc_wb(t, t->parameters);
     t->types = ftypes;
     if (ftypes != NULL) jl_gc_wb(t, t->types);
-    t->size = 0;
 
     t->name = NULL;
     if (jl_is_typename(name)) {
@@ -786,9 +786,12 @@ JL_DLLEXPORT jl_datatype_t *jl_new_primitivetype(jl_value_t *name, jl_module_t *
     uint32_t alignm = next_power_of_two(nbytes);
     if (alignm > MAX_ALIGN)
         alignm = MAX_ALIGN;
+    // memoize isprimitivetype, since it is much easier than checking
+    // (dta->name->names == svec() && dta->layout && dta->layout->size != 0)
+    // and we easily have a free bit for it in the DataType flags
+    bt->isprimitivetype = 1;
     bt->isbitstype = (parameters == jl_emptysvec);
-    bt->size = nbytes;
-    bt->layout = jl_get_layout(0, 0, alignm, 0, NULL, NULL);
+    bt->layout = jl_get_layout(nbytes, 0, 0, alignm, 0, NULL, NULL);
     bt->instance = NULL;
     return bt;
 }
@@ -803,10 +806,10 @@ JL_DLLEXPORT jl_datatype_t * jl_new_foreign_type(jl_sym_t *name,
 {
     jl_datatype_t *bt = jl_new_datatype(name, module, super,
       jl_emptysvec, jl_emptysvec, jl_emptysvec, jl_emptysvec, 0, 1, 0);
-    bt->size = large ? GC_MAX_SZCLASS+1 : 0;
     jl_datatype_layout_t *layout = (jl_datatype_layout_t *)
       jl_gc_perm_alloc(sizeof(jl_datatype_layout_t) + sizeof(jl_fielddescdyn_t),
         0, 4, 0);
+    layout->size = large ? GC_MAX_SZCLASS+1 : 0;
     layout->nfields = 0;
     layout->alignment = sizeof(void *);
     layout->haspadding = 1;
@@ -1065,7 +1068,7 @@ JL_DLLEXPORT jl_value_t *jl_atomic_cmpswap_bits(jl_datatype_t *dt, jl_datatype_t
     // n.b.: this does not spuriously fail if there are padding bits
     jl_task_t *ct = jl_current_task;
     int isptr = jl_field_isptr(rettyp, 0);
-    jl_value_t *y = jl_gc_alloc(ct->ptls, isptr ? nb : rettyp->size, isptr ? dt : rettyp);
+    jl_value_t *y = jl_gc_alloc(ct->ptls, isptr ? nb : jl_datatype_size(rettyp), isptr ? dt : rettyp);
     int success;
     jl_datatype_t *et = (jl_datatype_t*)jl_typeof(expected);
     if (nb == 0) {
@@ -1154,7 +1157,7 @@ JL_DLLEXPORT jl_value_t *jl_atomic_cmpswap_bits(jl_datatype_t *dt, jl_datatype_t
     }
     if (isptr) {
         JL_GC_PUSH1(&y);
-        jl_value_t *z = jl_gc_alloc(ct->ptls, rettyp->size, rettyp);
+        jl_value_t *z = jl_gc_alloc(ct->ptls, jl_datatype_size(rettyp), rettyp);
         *(jl_value_t**)z = y;
         JL_GC_POP();
         y = z;
@@ -1178,8 +1181,7 @@ JL_DLLEXPORT jl_value_t *jl_typemax_uint(jl_value_t *bt)
 
 #define PERMBOXN_FUNC(nb,nw)                                            \
     jl_value_t *jl_permbox##nb(jl_datatype_t *t, int##nb##_t x)         \
-    {   /* NOTE: t must be a concrete isbits datatype */                \
-        assert(jl_datatype_size(t) == sizeof(x));                       \
+    {   /* n.b. t must be a concrete isbits datatype of the right size */ \
         jl_value_t *v = jl_gc_permobj(nw * sizeof(void*), t);           \
         *(int##nb##_t*)jl_data_ptr(v) = x;                              \
         return v;                                                       \
@@ -1556,6 +1558,7 @@ static inline void memassign_safe(int hasptr, jl_value_t *parent, char *dst, con
         memmove_refs((void**)dst, (void**)src, nptr);
         jl_gc_multi_wb(parent, src);
         src = (jl_value_t*)((char*)src + nptr * sizeof(void*));
+        dst = dst + nptr * sizeof(void*);
         nb -= nptr * sizeof(void*);
     }
     else {
@@ -1824,7 +1827,7 @@ jl_value_t *replace_nth_field(jl_datatype_t *st, jl_value_t *v, size_t i, jl_val
                 rty = jl_nth_union_component(rty, *psel);
             }
             assert(!jl_field_isptr(rettyp, 0));
-            r = jl_gc_alloc(ct->ptls, rettyp->size, (jl_value_t*)rettyp);
+            r = jl_gc_alloc(ct->ptls, jl_datatype_size(rettyp), (jl_value_t*)rettyp);
             int success = (rty == jl_typeof(expected));
             if (needlock)
                 jl_lock_value(v);