+++ src/llvm-multiversioning.cpp
@@ -281,7 +281,7 @@ private:
     function_ref<LoopInfo&(Function&)> GetLI;
     function_ref<CallGraph&()> GetCG;
 
-    // Map from original functiton to one based index in `fvars`
+    // Map from original function to one based index in `fvars`
     std::map<const Function*,uint32_t> func_ids{};
     std::vector<Function*> orig_funcs{};
     std::vector<uint32_t> func_infos{};
@@ -1134,8 +1134,9 @@ static bool runMultiVersioning(Module &M, function_ref<LoopInfo&(Function&)> Get
     // At this point, we should have fixed up all the uses of the cloned functions
     // and collected all the shared/target-specific relocations.
     clone.emit_metadata();
-
+#ifdef JL_VERIFY_PASSES
     assert(!verifyModule(M, &errs()));
+#endif
 
     return true;
 }