+++ NEWS.md
@@ -9,6 +9,7 @@ New language features
   handled via `Base.split_rest`. ([#42902])
 * Character literals now support the same syntax allowed in string literals; i.e. the syntax can
   represent invalid UTF-8 sequences as allowed by the `Char` type ([#44989]).
+* Support for Unicode 15 ([#47392]).
 * Nested combinations of tuples and named tuples of symbols are now allowed as type parameters ([#46300]).
 
 Language changes
@@ -78,6 +79,8 @@ New library functions
 * New function `stack(x)` which generalises `reduce(hcat, x::Vector{<:Vector})` to any dimensionality,
   and allows any iterators of iterators. Method `stack(f, x)` generalises `mapreduce(f, hcat, x)` and
   is efficient. ([#43334])
+* New macro `@allocations` which is similar to `@allocated` except reporting the total number of allocations
+  rather than the total size of memory allocated. ([#47367])
 
 Library changes
 ---------------
@@ -94,6 +97,7 @@ Library changes
   a `Slices` object, which allows dispatching to provide more efficient methods ([#32310]).
 * `@kwdef` is now exported and added to the public API ([#46273])
 * An issue with order of operations in `fld1` is now fixed ([#28973]).
+* Sorting is now always stable by default as `QuickSort` was stabilized in ([#45222]).
 
 Standard library changes
 ------------------------